refactor: main page 를 index 로 옮기기
